Delete /data/db/mongod.lock only if exists

+++ docker-entrypoint.sh
@@ -8,7 +8,9 @@ trap 'kill -2 1; wait 1' SIGTERM
 #     exec /usr/bin/mongod "$@"
 # fi
 
-rm /data/db/mongod.lock
-mongod --repair
+if [ -e /data/db/mongod.lock ]; then
+	rm /data/db/mongod.lock
+	mongod --repair
+fi
 
 exec "$@"
